+++ tests/api/test-to-fixed-dynamic-buffer.c
@@ -57,6 +57,21 @@ final top: 1
 ==> rc=1, result='Error: invalid index'
 *** test_5b (duk_safe_call)
 ==> rc=1, result='Error: invalid index'
+*** test_6a (duk_safe_call)
+sz=16
+p[0]=123, buf[0]=0
+final top: 0
+==> rc=0, result='undefined'
+*** test_6b (duk_safe_call)
+sz=16
+p[0]=123, buf[0]=0
+final top: 0
+==> rc=0, result='undefined'
+*** test_6c (duk_safe_call)
+sz=16
+p[0]=123, buf[0]=123
+final top: 0
+==> rc=0, result='undefined'
 ===*/
 
 static void dump_buffer(duk_context *ctx) {
@@ -272,6 +287,87 @@ static duk_ret_t test_5b(duk_context *ctx) {
 	return 0;
 }
 
+/* When converting from an external buffer to a fixed buffer a copy is
+ * always made.
+ */
+static duk_ret_t test_6a(duk_context *ctx) {
+	unsigned char buf[16];
+	int i;
+	unsigned char *p;
+	duk_size_t sz;
+
+	for (i = 0; i < 16; i++) {
+		buf[i] = (unsigned char) i;
+	}
+
+	duk_push_external_buffer(ctx);
+	duk_config_buffer(ctx, -1, (void *) buf, 16);
+
+	p = (unsigned char *) duk_to_fixed_buffer(ctx, -1, &sz);
+	printf("sz=%ld\n", (long) sz); fflush(stdout);
+	p[0] = (unsigned char) 123;
+	printf("p[0]=%u, buf[0]=%u\n", (unsigned int) p[0], (unsigned int) buf[0]);
+
+	duk_pop(ctx);
+
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
+	return 0;
+}
+
+/* When converting from an external buffer to a dynamic buffer a copy
+ * is always made.
+ */
+static duk_ret_t test_6b(duk_context *ctx) {
+	unsigned char buf[16];
+	int i;
+	unsigned char *p;
+	duk_size_t sz;
+
+	for (i = 0; i < 16; i++) {
+		buf[i] = (unsigned char) i;
+	}
+
+	duk_push_external_buffer(ctx);
+	duk_config_buffer(ctx, -1, (void *) buf, 16);
+
+	p = (unsigned char *) duk_to_dynamic_buffer(ctx, -1, &sz);
+	printf("sz=%ld\n", (long) sz); fflush(stdout);
+	p[0] = (unsigned char) 123;
+	printf("p[0]=%u, buf[0]=%u\n", (unsigned int) p[0], (unsigned int) buf[0]);
+
+	duk_pop(ctx);
+
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
+	return 0;
+}
+
+/* When converting from an external buffer to a "don't care" buffer,
+ * an external buffer is kept as is.
+ */
+static duk_ret_t test_6c(duk_context *ctx) {
+	unsigned char buf[16];
+	int i;
+	unsigned char *p;
+	duk_size_t sz;
+
+	for (i = 0; i < 16; i++) {
+		buf[i] = (unsigned char) i;
+	}
+
+	duk_push_external_buffer(ctx);
+	duk_config_buffer(ctx, -1, (void *) buf, 16);
+
+	p = (unsigned char *) duk_to_buffer(ctx, -1, &sz);
+	printf("sz=%ld\n", (long) sz); fflush(stdout);
+	p[0] = (unsigned char) 123;
+	printf("p[0]=%u, buf[0]=%u\n", (unsigned int) p[0], (unsigned int) buf[0]);
+
+	duk_pop(ctx);
+
+	printf("final top: %ld\n", (long) duk_get_top(ctx));
+	return 0;
+}
+
 void test(duk_context *ctx) {
 	TEST_SAFE_CALL(test_1a);
 	TEST_SAFE_CALL(test_1b);
@@ -283,4 +379,7 @@ void test(duk_context *ctx) {
 	TEST_SAFE_CALL(test_4b);
 	TEST_SAFE_CALL(test_5a);
 	TEST_SAFE_CALL(test_5b);
+	TEST_SAFE_CALL(test_6a);
+	TEST_SAFE_CALL(test_6b);
+	TEST_SAFE_CALL(test_6c);
 }